fix: move authOptions to lib and fix type errors
- Move authOptions to app/lib/auth.ts (Next.js 15 route exports)
- Update imports in route and jobs actions
- Fix JobStatus type casting in jobs page
