+++ app/admin/golem/jobs/actions/jobs.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/lib/auth';
 import { createClient } from '@/lib/supabase/server';
 
 // Verify session before any operation
